requirements.txt
GitPython==3.1.44
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
python_dateutil==2.9.0.post0
seaborn==0.13.2
scipy==1.15.2
